pyproject.toml
# Poetry pyproject.toml: https://python-poetry.org/docs/pyproject/

[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "songmam"
version = "1.7.2"
description = "a facebook messenger hypermodern python library based on fastapi. "
readme = "README.md"
authors = [
  "codustry <hello@codustry.com>"
]
license = "MIT"
repository = "https://github.com/codustry/songmam"
homepage = "https://github.com/codustry/songmam"

# Keywords description https://python-poetry.org/docs/pyproject/#keywords
keywords = []  # Update me

# Pypi classifiers: https://pypi.org/classifiers/
classifiers = [  # Update me
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Topic :: Software Development :: Libraries :: Python Modules",
]

[tool.poetry.scripts]
# Entry points for the package https://python-poetry.org/docs/pyproject/#scripts
"songmam" = "songmam.__main__:app"

[tool.poetry.dependencies]
python = "^3.8"
loguru = "^0.5.1"
httpx = "^0.13.3"
furl = "^2.1.0"
avajana = "^0.4"
autoname = "^0.1.2"
parse = "^1.16.0"
path = "^15.0.0"
moshimoshi = "^0.2"
arrow = "^0.17.0"
fastapi = "^0.61.2"

[tool.poetry.dev-dependencies]
darglint = "^1.5.4"
isort = "^5.5.4"
pyupgrade = "^2.7.2"
black = "^20.8b1"
mypy = "^0.790"
bandit = "^1.6.2"
safety = "^1.9.0"
pytest = "^6.0.1"
pylint = "^2.5.3"
pydocstyle = "^5.0.2"
pre-commit = "^2.7.1"
faker = "^4.1.1"
pytest-asyncio = "^0.14.0"
python-decouple = "^3.3"
freezegun = "^1.0.0"

[tool.black]
# https://github.com/psf/black
line-length = 80
target-version = ["py38"]

[tool.isort]
# https://github.com/timothycrosley/isort/
known_typing = "typing,types,typing_extensions,mypy,mypy_extensions"
sections = "FUTURE,TYPING,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
include_trailing_comma = true
default_section = "FIRSTPARTY"
multi_line_output = 3
indent = 4
force_grid_wrap = 0
use_parentheses = true
line_length = 80
